+++ apps/frontend/src/libs/components/header/libs/components/notifications-popover/libs/components/notification-item/notification-item.tsx
@@ -1,0 +1,17 @@
+import styles from "./styles.module.css";
+
+type Properties = {
+	message: string;
+	timestamp: string;
+};
+
+const NotificationItem = ({ message, timestamp }: Properties): JSX.Element => {
+	return (
+		<div className={styles["notification-item"]}>
+			<p className={styles["notification-item-timestamp"]}>{timestamp}</p>
+			<p className={styles["notification-item-message"]}>{message}</p>
+		</div>
+	);
+};
+
+export { NotificationItem };