+++ apps/frontend/src/modules/notifications/libs/types/types.ts
@@ -1,0 +1,4 @@
+export {
+	type NotificationGetAllItemResponseDto,
+	type NotificationGetAllResponseDto,
+} from "@git-fit/shared";