+++ apps/frontend/src/pages/access-management/access-management.tsx
@@ -3,15 +3,18 @@ import {
 	ConfirmationModal,
 	Modal,
 	PageLayout,
+	Search,
 } from "~/libs/components/components.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
+	useAppForm,
 	useAppSelector,
 	useCallback,
 	useEffect,
 	useModal,
 	usePagination,
+	useSearch,
 	useState,
 } from "~/libs/hooks/hooks.js";
 import {
@@ -35,6 +38,7 @@ const AccessManagement = (): JSX.Element => {
 
 	const {
 		dataStatus: usersDataStatus,
+		deleteStatus: userDeleteStatus,
 		users,
 		usersTotalCount,
 	} = useAppSelector(({ users }) => users);
@@ -47,6 +51,11 @@ const AccessManagement = (): JSX.Element => {
 		groupsTotalCount,
 		groupUpdateStatus,
 	} = useAppSelector(({ groups }) => groups);
+
+	const { onSearch: onUserSearch, search: userSearch } = useSearch();
+	const { control, errors } = useAppForm({
+		defaultValues: { userSearch },
+	});
 
 	const {
 		onPageChange: onUserPageChange,
@@ -69,28 +78,38 @@ const AccessManagement = (): JSX.Element => {
 	});
 
 	const {
-		isOpened: isCreateModalOpened,
-		onClose: onCreateModalClose,
-		onOpen: onCreateModalOpen,
+		isOpened: isUserDeleteModalOpen,
+		onClose: onUserDeleteModalClose,
+		onOpen: onUserDeleteModalOpen,
 	} = useModal();
 
 	const {
-		isOpened: isUpdateModalOpened,
-		onClose: onUpdateModalClose,
-		onOpen: onUpdateModalOpen,
+		isOpened: isGroupCreateModalOpened,
+		onClose: onGroupCreateModalClose,
+		onOpen: onGroupCreateModalOpen,
 	} = useModal();
 
 	const {
-		isOpened: isDeleteModalOpen,
-		onClose: onDeleteModalClose,
-		onOpen: onDeleteModalOpen,
+		isOpened: isGroupUpdateModalOpened,
+		onClose: onGroupUpdateModalClose,
+		onOpen: onGroupUpdateModalOpen,
+	} = useModal();
+
+	const {
+		isOpened: isGroupDeleteModalOpen,
+		onClose: onGroupDeleteModalClose,
+		onOpen: onGroupDeleteModalOpen,
 	} = useModal();
 
 	const handleLoadUsers = useCallback(() => {
 		void dispatch(
-			userActions.loadAll({ page: userPage, pageSize: userPageSize }),
+			userActions.loadAll({
+				name: userSearch,
+				page: userPage,
+				pageSize: userPageSize,
+			}),
 		);
-	}, [dispatch, userPage, userPageSize]);
+	}, [dispatch, userPage, userPageSize, userSearch]);
 
 	const handleLoadGroups = useCallback(() => {
 		void dispatch(
@@ -117,6 +136,23 @@ const AccessManagement = (): JSX.Element => {
 		useState<GroupGetAllItemResponseDto | null>(null);
 	const hasGroupToEdit = groupToEdit !== null;
 
+	const [userToDeleteId, setUserToDeleteId] = useState<null | number>(null);
+	const hasUserToDelete = Boolean(userToDeleteId);
+
+	const handleUserDelete = useCallback(
+		(id: number): void => {
+			setUserToDeleteId(id);
+			onUserDeleteModalOpen();
+		},
+		[onUserDeleteModalOpen],
+	);
+
+	const handleUserDeleteSubmit = useCallback((): void => {
+		if (hasUserToDelete) {
+			void dispatch(userActions.deleteById({ id: userToDeleteId as number }));
+		}
+	}, [hasUserToDelete, dispatch, userToDeleteId]);
+
 	const [groupToDeleteId, setGroupToDeleteId] = useState<null | number>(null);
 	const hasGroupToDelete = Boolean(groupToDeleteId);
 
@@ -133,47 +169,55 @@ const AccessManagement = (): JSX.Element => {
 		}
 	}, [hasGroupToDelete, dispatch, groupToDeleteId]);
 
-	const handleEdit = useCallback(
+	const handleGroupEdit = useCallback(
 		(group: GroupGetAllItemResponseDto): void => {
 			setGroupToEdit(group);
-			onUpdateModalOpen();
+			onGroupUpdateModalOpen();
 		},
-		[onUpdateModalOpen],
+		[onGroupUpdateModalOpen],
 	);
 
-	const handleDelete = useCallback(
+	const handleGroupDelete = useCallback(
 		(id: number): void => {
 			setGroupToDeleteId(id);
-			onDeleteModalOpen();
+			onGroupDeleteModalOpen();
 		},
-		[onDeleteModalOpen],
+		[onGroupDeleteModalOpen],
 	);
+
+	useEffect(() => {
+		if (userDeleteStatus === DataStatus.FULFILLED) {
+			handleLoadUsers();
+			onUserDeleteModalClose();
+			setUserToDeleteId(null);
+		}
+	}, [userDeleteStatus, onUserDeleteModalClose, handleLoadUsers]);
 
 	useEffect(() => {
 		if (groupCreateStatus === DataStatus.FULFILLED) {
 			handleLoadUsers();
-			onCreateModalClose();
+			onGroupCreateModalClose();
 		}
-	}, [groupCreateStatus, onCreateModalClose, handleLoadUsers]);
+	}, [groupCreateStatus, onGroupCreateModalClose, handleLoadUsers]);
 
 	useEffect(() => {
 		if (groupUpdateStatus === DataStatus.FULFILLED) {
 			handleLoadUsers();
-			onUpdateModalClose();
+			onGroupUpdateModalClose();
 			setGroupToEdit(null);
 		}
-	}, [groupUpdateStatus, onUpdateModalClose, handleLoadUsers]);
+	}, [groupUpdateStatus, onGroupUpdateModalClose, handleLoadUsers]);
 
 	useEffect(() => {
 		if (groupDeleteStatus === DataStatus.FULFILLED) {
 			handleLoadUsers();
 			handleLoadGroups();
-			onDeleteModalClose();
+			onGroupDeleteModalClose();
 			setGroupToDeleteId(null);
 		}
 	}, [
 		groupDeleteStatus,
-		onDeleteModalClose,
+		onGroupDeleteModalClose,
 		handleLoadUsers,
 		handleLoadGroups,
 	]);
@@ -193,8 +237,20 @@ const AccessManagement = (): JSX.Element => {
 				<div className={styles["section-header"]}>
 					<h2 className={styles["section-title"]}>Users</h2>
 				</div>
+				<div className={styles["search-container"]}>
+					<Search
+						control={control}
+						errors={errors}
+						isLabelHidden
+						label="Users search"
+						name="userSearch"
+						onChange={onUserSearch}
+						placeholder="Enter name"
+					/>
+				</div>
 				<UsersTable
 					isLoading={isLoadingUsersData}
+					onDelete={handleUserDelete}
 					onPageChange={onUserPageChange}
 					onPageSizeChange={onUserPageSizeChange}
 					page={userPage}
@@ -208,15 +264,15 @@ const AccessManagement = (): JSX.Element => {
 				<div className={styles["section-header"]}>
 					<h2 className={styles["section-title"]}>Groups</h2>
 					<div>
-						<Button label="Create New" onClick={onCreateModalOpen} />
+						<Button label="Create New" onClick={onGroupCreateModalOpen} />
 					</div>
 				</div>
 
 				<GroupsTable
 					groups={groups}
 					isLoading={isLoadingGroupsData}
-					onDelete={handleDelete}
-					onEdit={handleEdit}
+					onDelete={handleGroupDelete}
+					onEdit={handleGroupEdit}
 					onPageChange={onGroupPageChange}
 					onPageSizeChange={onGroupPageSizeChange}
 					page={groupPage}
@@ -226,8 +282,8 @@ const AccessManagement = (): JSX.Element => {
 			</section>
 
 			<Modal
-				isOpened={isCreateModalOpened}
-				onClose={onCreateModalClose}
+				isOpened={isGroupCreateModalOpened}
+				onClose={onGroupCreateModalClose}
 				title="Create new group"
 			>
 				<GroupCreateForm onSubmit={handleGroupCreateSubmit} />
@@ -235,8 +291,8 @@ const AccessManagement = (): JSX.Element => {
 
 			{hasGroupToEdit && (
 				<Modal
-					isOpened={isUpdateModalOpened}
-					onClose={onUpdateModalClose}
+					isOpened={isGroupUpdateModalOpened}
+					onClose={onGroupUpdateModalClose}
 					title="Update group"
 				>
 					<GroupUpdateForm
@@ -249,9 +305,18 @@ const AccessManagement = (): JSX.Element => {
 			{hasGroupToDelete && (
 				<ConfirmationModal
 					content="The group will be deleted. This action cannot be undone. Do you want to continue?"
-					isOpened={isDeleteModalOpen}
-					onClose={onDeleteModalClose}
+					isOpened={isGroupDeleteModalOpen}
+					onClose={onGroupDeleteModalClose}
 					onConfirm={handleGroupDeleteSubmit}
+				/>
+			)}
+
+			{hasUserToDelete && (
+				<ConfirmationModal
+					content="The user will be deleted. This action cannot be undone. Do you want to continue?"
+					isOpened={isUserDeleteModalOpen}
+					onClose={onUserDeleteModalClose}
+					onConfirm={handleUserDeleteSubmit}
 				/>
 			)}
 		</PageLayout>