+++ apps/frontend/src/pages/project-access-management/libs/helpers/get-user-rows/get-user-rows.helper.ts
@@ -1,0 +1,13 @@
+import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
+
+import { type UserRow } from "../../types/types.js";
+
+const getUserRows = (users: UserGetAllItemResponseDto[]): UserRow[] =>
+	users.map((user) => ({
+		createdAt: user.createdAt,
+		id: user.id,
+		name: user.name,
+		projectGroups: user.projectGroups.map((projectGroup) => projectGroup.name),
+	}));
+
+export { getUserRows };