+++ packages/shared/src/index.ts
@@ -5,6 +5,7 @@ export {
 	ContentType,
 	ExceptionMessage,
 	PermissionKey,
+	ProjectPermissionKey,
 	ServerErrorType,
 	SortType,
 } from "./libs/enums/enums.js";
@@ -18,6 +19,7 @@ export {
 	checkHasPermission,
 	configureString,
 	formatDate,
+	formatRelativeTime,
 	getDifferenceInDays,
 	getRelativeDate,
 	getStartOfDay,
@@ -82,6 +84,14 @@ export {
 	type GroupUpdateResponseDto,
 	groupUpdateValidationSchema,
 } from "./modules/groups/groups.js";
+export {
+	type NotificationBulkCreateRequestDto,
+	type NotificationBulkCreateResponseDto,
+	type NotificationCreateRequestDto,
+	type NotificationGetAllItemResponseDto,
+	type NotificationGetAllResponseDto,
+	NotificationsApiPath,
+} from "./modules/notifications/notifications.js";
 export {
 	type PermissionGetAllItemResponseDto,
 	type PermissionGetAllResponseDto,